Add 'autodetect' functionality
This commit incorporates quite some changes. Most of these
changes were made to facilitate the 'autodetect' functionality
that is present with the Firmata protocol (termed "capability
query", and "analog mapping").

For autodetect functionality, the 'Port' and 'Pin' classes were
rewritten:

Pin:	a series of Boolean variables are added, which store
	the capabilities of that specific pin. The 'capability
	query response' modifies these.
Port:	not only the digital pins are added to Port, but also
	the analog pins (because Analog pins can be used as
	digital pin as well).
board:	The 'board' instance stores 'pins' and 'ports' as
	dict, with the number as key. To access pin 13,
	board.pins[13] gives the 13th instance.

One function was added: autodetect_layout() is the default,
and runs the 'capability query' tests. The existing function
is renamed to 'prefedined_layout()'.