Add 'capability query' (part 2)
This is the second round of modifications to make the
'capability query' functionality work.

The main modifications in this commit are focused on the
Analog pins, and interpretation of the 'analog_message' handle.
In short, the Pins are named after their digital number, yet
the Analog_message sent by Firmata gives the Analog number of
a Pin. Therefore, some kind of mapping between 'analog' and
'digital' number of the same pin was made. This is now done
with a new dict, 'board.map_analog_to_digital', which stores
the relation between Analog and Digital numbers. It is hackish
but works for now.

In addition, the 'enable/disable reporting' functions were
examined and updated if needed.

Lastly, the Pin.write() and Pin.read() functions were modified.
Now, they are proxies to a command from the Board instance.
(In my view, it is more elegant if the 'board' instance does
the writing and reading, as the 'pin' instance is a rather
dumb object (why should a Pin know how to read/write data!?).

(NB A more elegant solution would be to prepare commands in
a Queue object and let it be processed by the Board instance,
but this is for much later...)