Correct error: wrong 'write pin mode' command.
When setting the mode for an individual pin with the
Pin._set_mode() function, I accidentally swapped the
order of pin number and pin state. This resulted in
Firmata returning 'Unknown Pin Mode' as seen with the
Wireshark sniffer.

Reading the Firmata protocol, the command should be
[0xF4, pin nr, pin mode].

In addition, corrected two little errors in the 'enable/
disable reporting' function...