Add 'set sampling interval' function
From the Firmata protocol:
	"The sampling interval sets how often analog data
	and i2c data is reported to the client. The default
	value is 19 milliseconds."

The sampling_interval time can be set to 1 sec (1000 ms)
with the command: 'board.set_sampling_interval(1000)'

The result of this function can be examined by setting an
analog pin to 'enable_reporting', and then see the times
at which the values are reported in a USB sniffer such as
Wireshark. (Note Wireshark can detect USB communications
only as root under linux -- see its documentation).