Changed how Analog pins are stored on the board.
Previously, a dictionary was created (map_analog_to_digital)
in which key=Analog_pin_number, value=Digital_pin_number.
However, picking the correct analog pin was not straightforward,
because the digital pin number was required to obtain the
analog pins.

The 'map_analog_to_digital' is removed, and replaced with a
new dict, 'analog_pins', with keys=analog_pin_number, and
value holds a reference to the pin instance (not the pin #).

Old: board.pins[14]	        # on Arduino Due, this is A0 pin.
New: board.analog_pins[0]   # Again the A0 pin, easier to pick now

and these are equivalent now:
     board.pins[14]
     board.analog_pins[0]
     id(board.pins[14]) == id(board.analog_pins[0]) # True