Remove index reporting for firmata_version.
Sometimes -- for example, when a program is run twice without
disconnecting the USB port -- the firmata version number is
not reported. The board.firmata_version variable then is None.
If this is the case, the board.report_board_information()
function would raise a KeyError, since NoneType does not allow
for indexing (which would work well if version was '2.2' or
the like)....

There are no practical consequences linked to this commit.

TODO: obtain explicit firmware version/name in 'autodetect' mode